Restructure Repo using overlays and minor changes
Restructuring the repository using overlays for development and production environment. Further, a patch to the loadbalancer of the nginx-ingress is applied, which is necessary for our provider. Code which appears again in the ConfigMaps is moved to folder resources. ConfigMaps for web-Container are unified, such that it everything connected to the surface of the web container can now be found in this ConfigMap.