Initial setup (#2)
* Initial commit

* Add yamls for Jitsi setup

* Add second JVB deployment

* Update second JVB port to the same as first one

* Changed jvb-deployment to StatefulSet

* Create stateful set with config map

* Add service per pod controller

* Add missing metacontroller dependencies

* Remove unncessary port definition

* SSL termination at ingress and customizations

* Kustomization file updated

* First draft of HPA

* Allow shutdown via API call

* Optimizations

* Attach graceful shutdown script as preStop hook

* Add podManagementPolicy parallel to JVB config

* Finalize graceful shutdown

* Restructure Repo using overlays and minor changes

Restructuring the repository using overlays for development and production environment. Further, a patch to the loadbalancer of the nginx-ingress is applied, which is necessary for our provider. Code which appears again in the ConfigMaps is moved to folder resources. ConfigMaps for web-Container are unified, such that it everything connected to the surface of the web container can now be found in this ConfigMap.

* Certificates in overlays as patch by patchesStrategicMerge not possible

Also clean bug in jitsi-ingess-patch. Delete staging certificates as they are not used anymore.

* Restructuring of base

service-per-pod.yaml now splitted into service-per-pod-deployment (now only metacontroller) and service-per-pod-decoratorcontroller (now only jitsi) as it contained resources from both namespaces.

* Bug Fix kustomize.yaml

* Add monitoring patches

* Add resource requests and limits for JVB

* Change to mostly local yaml files

* Bug Fix: Remove duplicate resources

Remove duplicate resources in jitsi/service-per-pod-decoratorcontroller, which are also appearing in metacontroller namespace

* Add docker tags for JVB and prometheus exporter

* Update gitignore

* Fix kustomizations and add volume for prometheus

* Update Dashboard

* Fix Jitsi docker tags

* Update kustomize structure

* Restructuring

* Remove old jitsi-secret.yaml

Co-authored-by: Max <maximilian.kertel@woodmark.de>